refactor: type exporting
