+++ src/utils/index.ts
@@ -2,3 +2,4 @@ export * from "./context";
 export * from "./contracts";
 export * from "./poll";
 export * from "./misc";
+export * from "./db";
